Add link to stable package
This project currently doesn't work and its last update was a year ago -- as such, it's fair to say that the package is abandoned and shouldn't be used. However, this repository is the top hit on Google for search terms like "spark streaming facebook" which will confuse people. Thus, I think it would be good to point people to a package that is actively developed and maintained.